NonlinearRHS.jl

Luna.NonlinearRHS.TransFree — Method

(t::TransFree)(nl, Eω, z)

Calculate the reciprocal-domain (ω-kx-ky-space) nonlinear response due to the field Eω and place the result in nl.

Luna.NonlinearRHS.TransFree — Method

TransFree(grid, xygrid, FT, responses, densityfun, normfun)

Construct a TransFree to calculate the reciprocal-domain nonlinear polarisation.

Arguments

grid::AbstractGrid : the grid used in the simulation

xygrid : the spatial grid (instances of Grid.FreeGrid)

FT::FFTW.Plan : the full 3D (t-y-x) Fourier transform for the oversampled time grid

responses : Tuple of response functions

densityfun : callable which returns the gas density as a function of z

normfun : normalisation factor as fctn of z, can be created via norm_free

Luna.NonlinearRHS.TransModal — Type

TransModal

Transform E(ω) -> Pₙₗ(ω) for modal fields.

Luna.NonlinearRHS.TransModal — Method

TransModal(grid, ts, FT, resp, densityfun, norm!; rtol=1e-3, atol=0.0, mfcn=300, full=false)

Construct a TransModal, transform E(ω) -> Pₙₗ(ω) for modal fields.

Arguments

grid::AbstractGrid : the grid used in the simulation

ts::Modes.ToSpace : pre-created ToSpace for conversion from modal fields to space

FT::FFTW.Plan : the time-frequency Fourier transform for the oversampled time grid

resp : Tuple of response functions

densityfun : callable which returns the gas density as a function of z

norm! : normalisation function as fctn of z, can be created via norm_modal

rtol::Float=1e-3 : relative tolerance on the HCubature integration

atol::Float=0.0 : absolute tolerance on the HCubature integration

mfcn::Int=300 : maximum number of function evaluations for one modal integration

full::Bool=false : if true, use full 2-D mode integral, if false, only do radial integral

Luna.NonlinearRHS.TransModeAvg — Type

TransModeAvg

Transform E(ω) -> Pₙₗ(ω) for mode-averaged single-mode propagation.

Luna.NonlinearRHS.TransRadial — Type

TransRadial

Transform E(ω) -> Pₙₗ(ω) for radially symetric free-space propagation

Luna.NonlinearRHS.TransRadial — Method

(t::TransRadial)(nl, Eω, z)

Calculate the reciprocal-domain (ω-k-space) nonlinear response due to the field Eω and place the result in nl

Luna.NonlinearRHS.TransRadial — Method

TransRadial(grid, HT, FT, responses, densityfun, normfun)

Construct a TransRadial to calculate the reciprocal-domain nonlinear polarisation.

Arguments

grid::AbstractGrid : the grid used in the simulation

HT::QDHT : the Hankel transform which defines the spatial grid

FT::FFTW.Plan : the time-frequency Fourier transform for the oversampled time grid

responses : Tuple of response functions

densityfun : callable which returns the gas density as a function of z

normfun : normalisation factor as fctn of z, can be created via norm_radial

Luna.NonlinearRHS.Et_to_Pt! — Method

Et_to_Pt!(Pt, Et, responses, density)

Accumulate responses induced by Et in Pt.

Luna.NonlinearRHS.const_norm_free — Method

const_norm_free(grid, xygrid, nfun)

Make function to return normalisation factor for 3D propagation without re-calculating at every step.

Luna.NonlinearRHS.const_norm_radial — Method

const_norm_radial(ω, q, nfun)

Make function to return normalisation factor for radial symmetry without re-calculating at every step.

Luna.NonlinearRHS.copy_scale! — Method

copy_scale!(dest, source, N, scale)

Copy first N elements from source to dest and simultaneously multiply by scale factor. For multi-dimensional dest and source, work along first axis.

Luna.NonlinearRHS.copy_scale_both! — Method

copy_scale_both!(dest::Vector, source::Vector, N, scale)

Copy first and last N elements from source to first and last N elements in dest and simultaneously multiply by scale factor. For multi-dimensional dest and source, work along first axis.

Luna.NonlinearRHS.norm_free — Method

norm_free(grid, xygrid, nfun)

Make function to return normalisation factor for 3D propagation.

Note

Here, nfun(ω; z) needs to take frequency ω and a keyword argument z.

Luna.NonlinearRHS.norm_modal — Method

norm_modal(grid; shock=true)

Normalisation function for modal propagation. If shock is false, the intrinsic frequency dependence of the nonlinear response is ignored, which turns off optical shock formation/ self-steepening.

Luna.NonlinearRHS.norm_radial — Method

norm_radial(ω, q, nfun)

Make function to return normalisation factor for radial symmetry.

Note

Here, nfun(ω; z) needs to take frequency ω and a keyword argument z.

Luna.NonlinearRHS.to_freq! — Method

to_freq!(Aω, Aωo, Ato, FTplan)

Transform oversampled A(t) to A(ω) on normal grid

Luna.NonlinearRHS.to_time! — Method

to_time!(Ato, Aω, Aωo, IFTplan)

Transform $A(ω)$ on normal grid to $A(t)$ on oversampled time grid.