RectModes.jl

Luna.RectModes.RectMode — Method

convenience constructor for non-constant core index

Luna.RectModes.RectMode — Method

convenience constructor assunming single gas filling and specified cladding

Luna.Modes.neff — Method

neff(m::RectMode, ω; z=0)

effective index of rectangular mode with dielectric core and arbitrary (metal or dielectric) cladding.

Adapted from Laakmann, K. D. & Steier, W. H. Waveguides: characteristic modes of hollow rectangular dielectric waveguides. Appl. Opt., AO 15, 1334–1340 (1976).

I had to re-derive the result in order to get the complex cladding index contribution to the real part of neff.