Tools.jl

Luna.Tools.Ld — Method

Get dispersion length

Luna.Tools.Lfiss — Method

Get fission length

Luna.Tools.Lnl — Method

Get nonlinear length

Luna.Tools.getN — Method

Get soliton order

Luna.Tools.getN0n0n2 — Method

Get linear and nonlinear refractive index and gas number density

Luna.Tools.getβ2 — Method

Get GVD coefficient

Luna.Tools.getγ — Method

Get nonlinear coefficient

Luna.Tools.params — Method

Soliton parameter collection

Luna.Tools.pressureRDW — Method

pressureRDW(a::Number, gas::Symbol, λ_target, λ0; Pmax=100, clad=:SiO2, kwargs...)

Calculate the phase-matching pressure for resonant dispersive wave (RDW) emission at λ_target in a capillary with core radius a filled with gas when pumping at λ0.

Luna.Tools.λRDW — Method

λRDW(m::Modes.AbstractMode, λ0; z=0, λlims=(100e-9, 0.9λ0))

Calculate the phase-matching wavelength for resonant dispersive wave (RDW) emission in the mode m when pumping at λ0.

This neglects the nonlinear contribution to the phase mismatch.

Luna.Tools.λRDW — Method

λRDW(a::Number, gas::Symbol, pressure, λ0; λlims=(100e-9, 0.9λ0), kwargs...)

Calculate the phase-matching wavelength for resonant dispersive wave (RDW) emission in a capillary with core radius a filled with gas at a certain pressure when pumping at λ0. Additional kwargs are passed onto Capillary.MarcatilliMode.

This neglects the nonlinear contribution to the phase mismatch.

Luna.Tools.τfw_to_τ0 — Method

Calculate 'natural' pulse width from FWHM