Antiresonant.jl

Luna.Antiresonant.ZeisbergerMode — Method

ZeisbergerMode(args...; wallthickness, kwargs...)

Create a capillary-like mode with the effective index given by eq. (15) in [1].

wallthickness (mandatory kwarg) sets the thickness of the anti-resonant struts and loss (optional, defaults to true) can be either a Bool (to switch on/off loss completely) or a Real (to up/down-scale the loss given by the model). Other kwargs are passed on to the constructor of a Capillary.MarcatiliMode.

[1] Zeisberger, M., Schmidt, M.A. Analytic model for the complex effective index of the leaky modes of tube-type anti-resonant hollow core fibers. Sci Rep 7, 11761 (2017). https://doi.org/10.1038/s41598-017-12234-5